requirements.txt
streamlit
pandas
python-dotenv
langchain
groq
duckduckgo-search
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
duckduckgo-search
ddgs
langchain-community
langchain-groq
langchain-core
openpyxl
werkzeug
faiss-cpu
sentence-transformers
pypdf
openai
google-genai
requests